Move //test/integration/ConsistencyTest to //test/integration/database/IsolationTest
